r1-core-refactor: O(1) append_response with last_assistant_index cache
This implements the O(1) append_response optimization from REVIEW.md P1.

Changes:
- Added last_assistant_index field to AppState for O(1) message lookup
- Updated append_response to use cached index before falling back to search
- Updated index maintenance in reorder_agent_after_tools and move_turn_complete_to_end
- Applied clippy --fix for unused imports and other warnings
- All 671+ tests pass

+++ crates/runie-core/src/session_jsonl.rs
@@ -103,7 +103,7 @@ impl JsonlReader {
     /// Read the next domain event, or Ok(None) on EOF.
     pub fn read_event(&mut self) -> Result<Option<DomainEvent>> {
         let mut line = String::new();
-        match self.reader.read_line(&mut line)? as usize {
+        match self.reader.read_line(&mut line)? {
             0 => Ok(None),
             _ => {
                 self.line_no += 1;
